Struct jitsi_xmpp_parsers::jingle::Jingle

pub struct Jingle {
    pub action: Action,
    pub initiator: Option<Jid>,
    pub responder: Option<Jid>,
    pub sid: SessionId,
    pub contents: Vec<Content>,
    pub reason: Option<ReasonElement>,
    pub group: Option<Group>,
    pub other: Vec<Element>,
}

The main Jingle container, to be included in an iq stanza.

Fields
action: Action

The action to execute on both ends.

initiator: Option<Jid>

Who the initiator is.

responder: Option<Jid>

Who the responder is.

sid: SessionId

Unique session identifier between two entities.

contents: Vec<Content>

A list of contents to be negotiated in this session.

reason: Option<ReasonElement>

An optional reason.

group: Option<Group>

An optional grouping.

other: Vec<Element>

Payloads to be included.

Implementations

impl Jingle

pub fn new(action: Action, sid: SessionId) -> Jingle

Create a new Jingle element.

pub fn with_initiator(self, initiator: Jid) -> Jingle

Set the initiator’s JID.

pub fn with_responder(self, responder: Jid) -> Jingle

Set the responder’s JID.

pub fn add_content(self, content: Content) -> Jingle

Add a content to this Jingle container.

pub fn set_reason(self, reason: ReasonElement) -> Jingle

Set the reason in this Jingle container.

pub fn set_group(self, group: Group) -> Jingle

Set the grouping in this Jingle container.